updating names to include "Async" for commands, queries and handlers. Renaming test classes after what they're testing
