+++ AllReadyApp/Web-App/AllReady/Areas/Admin/Features/Tasks/AssignTaskCommandAsync.cs
@@ -3,7 +3,7 @@ using MediatR;
 
 namespace AllReady.Areas.Admin.Features.Tasks
 {
-    public class AssignTaskCommand : IAsyncRequest
+    public class AssignTaskCommandAsync : IAsyncRequest
     {
         public int TaskId { get; set; }
         public List<string> UserIds { get; set; }
