+++ AllReadyApp/Web-App/AllReady/Areas/Admin/Features/Tasks/DeleteTaskCommandAsync.cs
@@ -2,7 +2,7 @@
 
 namespace AllReady.Areas.Admin.Features.Tasks
 {
-    public class DeleteTaskCommand : IAsyncRequest
+    public class DeleteTaskCommandAsync : IAsyncRequest
     {
         public int TaskId {get; set;}
     }
