+++ AllReadyApp/Web-App/AllReady/Areas/Admin/Features/Tasks/EditTaskCommandAsync.cs
@@ -3,7 +3,7 @@ using MediatR;
 
 namespace AllReady.Areas.Admin.Features.Tasks
 {
-    public class EditTaskCommand : IAsyncRequest<int>
+    public class EditTaskCommandAsync : IAsyncRequest<int>
     {
         public TaskEditModel Task {get; set;}
     }
