+++ AllReadyApp/Web-App/AllReady/Areas/Admin/Features/Tasks/MessageTaskVolunteersCommandAsync.cs
@@ -3,7 +3,7 @@ using MediatR;
 
 namespace AllReady.Areas.Admin.Features.Tasks
 {
-    public class MessageTaskVolunteersCommand : IAsyncRequest
+    public class MessageTaskVolunteersCommandAsync : IAsyncRequest
     {
         public MessageTaskVolunteersModel Model {get; set;}
     }
